# apply StartsWith optimizations to EndsWith