@adamsitnik 2019-10-01
/azp run