@adamsitnik 2019-10-02
/azp run