@adamsitnik 2019-10-03
/azp run