@tarekgh 2019-10-03
@adamsitnik could you please test the following 2 cases and ensure work as expected with your changes?

```C#
            Console.WriteLine($"{CultureInfo.InvariantCulture.CompareInfo.IsSuffix("A\u030A", "Å")}"); // print True
            Console.WriteLine($"{CultureInfo.GetCultureInfo("sk-SK").CompareInfo.IsSuffix("ch", "h")}"); // print False
```