@adamsitnik 2019-10-04
> could you please test the following 2 cases and ensure work as expected with your changes?

Sure! The output:

![obraz](https://user-images.githubusercontent.com/6011991/66190257-eb0f4580-e68b-11e9-9ab0-ecc751e67390.png)