@lukaaash 2020-07-16
Apparently, in .NET 5.0 Preview 6, `value.EndsWith("\0")` returns `true` for all strings (unlike all other .NET platforms). Looks like an unintended side effect or a bug.