@tarekgh 2020-07-16
@lukaaash This is not a bug. in .NET 5.0 we have started to use ICU which has slight collation behavior than Windows NLS. '\0' on doesn't have any sort weight. If you don't like this behavior, you may consider using Ordinal behavior instead.