@safern 2020-07-16
@lukaaash please refer to: https://docs.microsoft.com/en-us/dotnet/standard/globalization-localization/globalization-icu to get more details about the motivation of the change.