@lukaaash 2020-07-21
@tarekgh @safern Thanks for the explanation and for the link! (In our code where this did cause a problem, we should actually have been using Ordinal comparison anyway.)